Add group membership notification settings
